In the NOT expression parser, you forgot to account for how there are a total of THREE characters to skip when reading the NOT expression!!!
The NOT expression is !(<expr>), which means that '!', '(', and ')' should be skipped. Otherwise, if another function is calling NOT, the parser may be one character behind (instead of past the NOT expression).

For example, imagine a case with &(&(!(<FALSE expr>)), <FALSE expr>).

!(<FALSE expr>) will be true, but when going back to its & function, the reader is left off on the first closing parentheses ) . Then, the inner & will terminate, moving the reader to the second closing parentheses. At this point, the outer & will see that the reader is pointing to a closing parentheses and assume there are no other expressions to evaluate. (and since the only evaluated inner expression was true, the entire output is true. this is incorrect because the outer & function should be false due to the second inner expression being <FALSE>)

HOWEVER, in reality, the reader should be pointing to the comma , that occurs right after the inner & expression. If the outer & expression saw a comma on the reader, it would reach the second inner expression (which is <FALSE>, making the outer & expression output false). Thus proving that the reader being behind by ONE character (after the NOT evaluation) can prove fatal.

TL;DR - Be sure to evaluate/implement your expressions very carefully. Underestimating the NOT expression to be more simple led to its closing parentheses being forgotten!

ALSO, make sure you test your code more thoroughly! The rough draft (first submission) felt very confident, but you forgot to test such a scenario that would highlight the error in the parseNotExpr() function. So, be more patient and think more outside of the box. Good effort! (: